Roll check of tab display into use tabs hook
